% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{player}
\alias{player}
\title{Create a player object}
\usage{
player(name, position, college)
}
\arguments{
\item{name}{Player name}

\item{position}{Player position}

\item{college}{Player college}
}
\value{
An object of class "player"
}
\description{
Create a player object
}
