% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player_summary.R
\name{player_summary}
\alias{player_summary}
\title{Get a player from the combine dataset}
\usage{
player_summary(name)
}
\arguments{
\item{name}{A character string to search for (partial matches allowed)}
}
\value{
A data frame of matching players
}
\description{
This function searches the NFL Combine dataset for players whose names
partially match the input string.
}
