% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position_summary.R
\name{position_summary}
\alias{position_summary}
\title{Summary of combine drills by position}
\usage{
position_summary(position)
}
\arguments{
\item{position}{A position abbreviation (e.g., "WR", "QB", "LB")}
}
\value{
A data frame of position averages
}
\description{
Summary of combine drills by position
}
