% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_player.R
\name{print.player}
\alias{print.player}
\title{Print method for player objects}
\usage{
\method{print}{player}(x, ...)
}
\arguments{
\item{x}{A player object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print method for player objects
}
